tidied up the code, changed log messages
- changed all warning-inducing places in code so there's no compile
warnings anymore;
- added some new useful log messages;
- removed some useless log messages;
- changed some log messages;

+++ DF Sources/g_items.pas
@@ -284,7 +284,7 @@ begin
   begin
     find_id := ForcedID;
 
-    if find_id > High(gItems) then
+    if find_id > DWORD(High(gItems)) then
       SetLength(gItems, find_id + 32);
   end;
